Small Project/ConvertPDF/requirements.txt
Flask==3.0.0
img2pdf==0.5.1
Pillow==10.2.0
python-docx==1.1.0
reportlab==4.0.9
Werkzeug==3.0.1
gunicorn==21.2.0
